Move `asyncActionAsValue` alongside similar functions
